requirements.txt
boto3>=1.34.0
chainlit>=1.0.0
aws-cdk-lib>=2.110.0
constructs>=10.0.0
